less complexity on init, moved config to config.js
