+++ app/config.js
@@ -1,5 +1,9 @@
 var path = require('path')
 var os = require('os')
+var fs = require('fs')
+var app = require('app')
+
+var IPFS_PATH_FILE = app.getDataPath() + '/ipfs-electron-app-node-path'
 
 exports = module.exports = {
   'menu-bar-width': 240,
@@ -12,5 +16,16 @@ exports = module.exports = {
   'tray-icon': (os.platform() !== 'darwin' ?
     path.resolve(__dirname, '../node_modules/ipfs-logo/ipfs-logo-256-ice.png')
     : path.resolve(__dirname, '../node_modules/ipfs-logo/platform-icons/osx-menu-bar.png')),
-  'webui-path': '/webui'
+  'webui-path': '/webui',
+  'ipfs-path': (function () {
+    var pathIPFS
+    try {
+      pathIPFS = fs.readFileSync(IPFS_PATH_FILE, 'utf-8')
+    } catch (e) {
+      pathIPFS = process.env.IPFS_PATH ||
+       (process.env.HOME || process.env.USERPROFILE) + '/.ipfs'
+    }
+    return pathIPFS
+  })(),
+  'ipfs-path-file': IPFS_PATH_FILE
 }
